EJ. Made Major Directory Structure Changes..
