+++ .github/CODEOWNERS
@@ -19,9 +19,9 @@
 # subdirectories.
 /build/logs/ @EbenezerJesuraj
 
-# @sujairamprasathc owns any file in an apps directory
+# @EbenezerJesuraj owns any file in an apps directory
 # anywhere in your repository.
-apps/ @sujairamprasathc
+apps/ @EbenezerJesuraj
 
 # @sujairamprasathc owns any file in the `/docs`
 # directory in the root of your repository and any of its
