--- Jarvis' Polygon.txt
@@ -1,12 +1,0 @@
-106 229
-125 291
-179 297
-338 289
-396 267
-395 222
-393 155
-365 105
-270 102
-123 101
-118 111
-106 229
