--- Points.txt
@@ -1,99 +1,0 @@
-334 150
-
-142 176
-
-271 122
-
-385 265
-
-270 102
-
-303 282
-
-272 268
-
-396 267
-
-139 221
-
-365 105
-
-247 276
-
-294 256
-
-285 188
-
-395 222
-
-389 184
-
-334 214
-
-342 124
-
-246 228
-
-294 213
-
-116 232
-
-366 216
-
-184 271
-
-239 190
-
-264 233
-
-338 130
-
-106 229
-
-386 237
-
-125 253
-
-227 171
-
-207 172
-
-138 116
-
-140 141
-
-221 181
-
-246 187
-
-171 115
-
-232 114
-
-338 289
-
-379 268
-
-201 129
-
-159 143
-
-125 291
-
-329 165
-
-118 111
-
-334 206
-
-179 297
-
-114 241
-
-234 169
-
-302 154
-
-393 155
-
-123 101
