+++ src/Graham's Polygon.txt
@@ -1,0 +1,12 @@
+123 101
+270 102
+365 105
+393 155
+395 222
+396 267
+338 289
+179 297
+125 291
+106 229
+118 111
+123 101
