+++ src/Graham's Scan Convex Hull.bmp
