+++ src/Jarvis' March Convex Hull.bmp
